+++ memberaudit/models/general.py
@@ -127,6 +127,7 @@ class ComplianceGroupDesignation(models.Model):
 class Location(models.Model):
     """An Eve Online location: Station or Upwell Structure or Solar System"""
 
+    _ASSET_SAFETY_ID = 2004
     _SOLAR_SYSTEM_ID_START = 30_000_000
     _SOLAR_SYSTEM_ID_END = 33_000_000
     _STATION_ID_START = 60_000_000
@@ -227,6 +228,10 @@ class Location(models.Model):
     @classmethod
     def is_structure_id(cls, location_id: int) -> bool:
         return location_id >= cls._STRUCTURE_ID_START
+
+    @classmethod
+    def is_asset_safety_id(cls, location_id: int) -> bool:
+        return location_id == cls._ASSET_SAFETY_ID
 
 
 class EveShipType(EveType):